moved UI files
Moved actions, components, stores and session.js from js/ to top level.
