+++ actions/chatMessageActions.js
@@ -4,7 +4,7 @@
  * Sends a message to the server.
  */
 
-var Dispatcher = require('../uiDispatcher')
+var Dispatcher = require('../js/uiDispatcher')
 
 var session = require('../session')
