+++ components/main.js
